requirements.txt
langchain
langchain-community
google-genai
faiss-cpu
pypdf
fastapi
uvicorn
python-dotenv
